+++ webapp/src/app/shared/widgets/event-widget/event-editor/event-display.model.ts
@@ -1,0 +1,15 @@
+import { ChangeDetectorRef } from '@angular/core';
+import { AbstractEvent } from '../event-registry/abstract-event';
+
+export interface EventDisplay {
+	text: string;
+	draggable: boolean;
+    isActionStep: boolean;
+    isSelected: boolean;
+    parent: AbstractEvent<any>[];
+	data?: AbstractEvent<any>;
+    children?: AbstractEvent<any>[];
+	
+    level: number;
+    changeDetector?: ChangeDetectorRef;
+}