+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ## [Unreleased]
 ## [0.8.1] 2020-12-28
 
+### Added
+- Event editor rework
+
 ### Changed
 - Improved tile drawing by drawing a line of points instead of single points. Avoids holes when drawing fast. [#189](https://github.com/CCDirectLink/crosscode-map-editor/pull/189)
 - Added keyboard controls to list search overlay (add new entity / add new event)